Restructured the entire project
